Uploaded the Functions and Test Files
Uploaded the Functions Python file and the Test Python file.

The Functions python file has the actual functions that can be imported and used.

The Test file serves to be an example. Using multiple vectors and running those vectors on the measures in the Functions file. This file is not required to use the different measures that are in the other file.